spm to npm
