--- .spmignore
@@ -1,11 +1,0 @@
-.idea
-dist
-.gitattributes
-.gitignore
-.jshintrc
-.spmignore
-.travis.yml
-LICENSE-MIT
-spm_modules
-node_modules
-_site
