+++ example/index.html
@@ -1,0 +1,11 @@
+<!DOCTYPE html>
+<html>
+  <head>
+    <meta charset="utf-8">
+    <title>bootstrap3-grid</title>
+    <link href="http://ue.17173cdn.com/a/lib/gallery/bootstrap/3.0.0/bootstrap.css" rel="stylesheet">
+  </head>
+  <body>
+    <!-- built files will be auto injected -->
+  </body>
+</html>
