+++ example/main.js
@@ -1,39 +1,28 @@
-# bootstrap3-grid
+'use strict';
 
----
-源码来自[Simple-Paging-Grid](https://github.com/JamesRandall/Simple-Paging-Grid)，把它改为了CMD模块的
+var $ = require('jquery');
 
----
+require('../src/bootstrap3-grid.less');
+require('../src/bootstrap3-grid');
 
-[![Build Status](https://travis-ci.org/lynzz/bootstrap3-grid.svg?branch=master)](https://travis-ci.org/lynzz/bootstrap3-grid)
-[![Coverage Status](https://coveralls.io/repos/lynzz/bootstrap3-grid/badge.png)](https://coveralls.io/r/lynzz/bootstrap3-grid)
+$('<div id="exampleGrid"></div>').appendTo('body');
 
-## Install
-
-```
-$ npm install bootstrap3-grid --save
-```
-
-## Usage
-
-````html
-<div id="exampleGrid">
-</div>
-````
-
-```javascript
-require('bootstrap3-grid');
 $("#exampleGrid").simplePagingGrid({
     columnNames: ["OrderLineID", "Name", "Price ($)", "Quantity"],
     columnKeys: ["OrderLineID", "Name", "Price", "Quantity"],
-    data: [
-        { "OrderLineID": 1, "Name": "Pineapple", "Price": 1.50, "Quantity": 4 },
-        { "OrderLineID": 2, "Name": "Fresh Spring Onions", "Price": 1.10, "Quantity": 40 },
-        { "OrderLineID": 3, "Name": "Oranges", "Price": 0.20, "Quantity": 8 },
-        { "OrderLineID": 4, "Name": "Apples", "Price": 1.50, "Quantity": 5 },
-        { "OrderLineID": 5, "Name": "Raspberries", "Price": 1.50, "Quantity": 20 },
-        { "OrderLineID": 6, "Name": "Blueberries", "Price": 1.50, "Quantity": 20 },
-        { "OrderLineID": 7, "Name": "Pairs", "Price": 1.50, "Quantity": 8 },
+    pageNumber: 1,
+    data: null,
+    templates: {
+      emptyTemplate: 'Not rows'
+    },
+    data1: [
+        // { "OrderLineID": 1, "Name": "Pineapple", "Price": 1.50, "Quantity": 4 },
+        // { "OrderLineID": 2, "Name": "Fresh Spring Onions", "Price": 1.10, "Quantity": 40 },
+        // { "OrderLineID": 3, "Name": "Oranges", "Price": 0.20, "Quantity": 8 },
+        // { "OrderLineID": 4, "Name": "Apples", "Price": 1.50, "Quantity": 5 },
+        // { "OrderLineID": 5, "Name": "Raspberries", "Price": 1.50, "Quantity": 20 },
+        // { "OrderLineID": 6, "Name": "Blueberries", "Price": 1.50, "Quantity": 20 },
+        // { "OrderLineID": 7, "Name": "Pairs", "Price": 1.50, "Quantity": 8 },
         { "OrderLineID": 8, "Name": "Melons", "Price": 1.50, "Quantity": 2 },
         { "OrderLineID": 9, "Name": "Potatoes", "Price": 1.50, "Quantity": 6 },
         { "OrderLineID": 10, "Name": "Sweet Potatoes", "Price": 1.50, "Quantity": 3 },
@@ -50,5 +39,3 @@ $("#exampleGrid").simplePagingGrid({
         { "OrderLineID": 21, "Name": "Beetroot", "Price": 0.30, "Quantity": 3 },
         { "OrderLineID": 22, "Name": "Avocado", "Price": 2.30, "Quantity": 1 }]
 });
-
-```
