--- index.js
@@ -1,2 +1,0 @@
-module.exports = require('./src/bootstrap3-grid');
-
