+++ src/bootstrap3-grid.less
@@ -54,4 +54,3 @@ div.loading
     background-repeat: no-repeat;
     background-position: center;
 }
-
