+++ test/specs/check-spec.js
@@ -1,10 +1,9 @@
-var expect = require('expect.js');
 var $ = require('jquery');
 $(document.body).css({
     'margin': 0,
     'padding': 0
 });
-require('../src/bootstrap3-grid');
+require('../../src/bootstrap3-grid');
 
 var $gridElement;
 
@@ -33,5 +32,3 @@ function simpleData() {
         {Name: "Oranges", Price: 2.25}
     ];
 }
-
-
