+++ test/specs/data-url-binding-spec.js
@@ -1,12 +1,10 @@
-var expect = require('expect.js');
-var sinon = require('sinon');
 var $ = require('jquery');
 
 $(document.body).css({
     'margin': 0,
     'padding': 0
 });
-require('../src/bootstrap3-grid');
+require('../../src/bootstrap3-grid');
 
 var $gridElement;
 
@@ -59,5 +57,3 @@ function simpleData() {
         {Name: "Oranges", Price: 2.25}
     ];
 }
-
-
